Update man page
Setting KDUMP_CPUS to a very high value - for example, 900 or more can
cause the kdump/fadump kernel to fail during boot due to limited
system resources in the dump capture environment.

So Add a note in the man page to warn users about the potential
consequences of setting KDUMP_CPUS to a high value.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>